# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentação da API do sistema",
        "contact": {
            "email": "contato@exemplo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8989"
        }
    ],
    "paths": {
        "/api/advertisement/{id}": {
            "get": {
                "tags": [
                    "Advertisements"
                ],
                "summary": "Obtém uma propaganda específica",
                "description": "Retorna os detalhes de uma propaganda específica pelo ID",
                "operationId": "getAdvertisement",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da propaganda",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da propaganda",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Advertisement"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Propaganda não encontrada"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Advertisements"
                ],
                "summary": "Atualiza uma propaganda existente",
                "description": "Atualiza os dados de uma propaganda específica",
                "operationId": "updateAdvertisement",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da propaganda",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAdvertisementRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Propaganda atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Advertisement"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Propaganda não encontrada"
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Advertisements"
                ],
                "summary": "Remove uma propaganda",
                "description": "Remove uma propaganda específica do sistema",
                "operationId": "destroyAdvertisement",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da propaganda",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Propaganda removida com sucesso"
                    },
                    "404": {
                        "description": "Propaganda não encontrada"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/advertisement": {
            "get": {
                "tags": [
                    "Advertisements"
                ],
                "summary": "Lista todas as propagandas",
                "description": "Retorna uma lista de todas as propagandas cadastradas",
                "operationId": "indexAdvertisement",
                "responses": {
                    "200": {
                        "description": "Lista de propagandas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Advertisement"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Advertisements"
                ],
                "summary": "Cria uma nova propaganda",
                "description": "Cria uma nova propaganda com os dados fornecidos",
                "operationId": "storeAdvertisement",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreAdvertisementRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Propaganda criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Advertisement"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Advertisement": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Promoção de Verão"
                    },
                    "description": {
                        "type": "string",
                        "example": "Descontos especiais para o verão"
                    },
                    "image": {
                        "type": "string",
                        "example": "https://exemplo.com/imagem.jpg",
                        "nullable": true
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "promotion_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-03-03T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-03-03T00:00:00Z"
                    }
                },
                "type": "object"
            },
            "StoreAdvertisementRequest": {
                "required": [
                    "title",
                    "description"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Promoção de Verão"
                    },
                    "description": {
                        "type": "string",
                        "example": "Descontos especiais para o verão"
                    },
                    "image": {
                        "type": "string",
                        "example": "https://exemplo.com/imagem.jpg",
                        "nullable": true
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "promotion_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdateAdvertisementRequest": {
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Promoção de Verão"
                    },
                    "description": {
                        "type": "string",
                        "example": "Descontos especiais para o verão"
                    },
                    "image": {
                        "type": "string",
                        "example": "https://exemplo.com/imagem.jpg",
                        "nullable": true
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "promotion_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}